Reparo no cast int to float;

+++ app/aluno/lista_notas_detalhe.php
@@ -71,7 +71,7 @@ $quantidade_notas_diario = $conn->get_one($sql_quantidade_notas);
             	//Fim
 ?>
 
-<!--IniÃ‚Â­cio - Victor Uliisses Pugliese - 15:28 01/05/2012 - Tabela Media -->
+<!--Inicio - Victor Uliisses Pugliese - 15:28 01/05/2012 - Tabela Media -->
 <table style="width: 220px; font-size: 12px" >
 	<tr bgcolor="#EEEEEE">
 		<td><b>Média</b></td>
@@ -167,7 +167,7 @@ $quantidade_notas_diario = $conn->get_one($sql_quantidade_notas);
             		if($disciplina_aluno['nota'] != -1)
             		{
 	            		echo '<td align="center">'. number::numeric2decimal_br($disciplina_aluno['nota'],1) .'</td>';        
-	            		$media_aluno += (int) $disciplina_aluno['nota'];
+	            		$media_aluno += (float) $disciplina_aluno['nota'];
 	            	}
 	            	else
 	            		echo '<td align="center"> 0,0 </td>';
@@ -189,7 +189,7 @@ $quantidade_notas_diario = $conn->get_one($sql_quantidade_notas);
              	if($disciplina_aluno['nota_distribuida'] != -1)
 	           	{
 	           		echo '<td align="center">'. $disciplina_aluno['nota_distribuida'] .'</td>';        
-	           		$nota_disc_maxima += (int) $disciplina_aluno['nota_distribuida'];
+	           		$nota_disc_maxima += (float) $disciplina_aluno['nota_distribuida'];
 	           	}
 	           	else
 	           		echo '<td align="center"> 0,0 </td>';
